# Improved text wrapping performance