## What does the pull request do?

This PR optimizes text layout when wrapping is enabled.

There are two main changes:

 - CPU: avoid calling `FetchTextRuns` in `TextFormatterImpl.FormatLine` only to continue wrapping previously shaped runs. The results were already completely discarded in this case: let's completely avoid processing instead. This saves a **lot** of CPU time. The method code has been reorganized to make that path clearer.

- Memory: made `LineBreak.RemainingRuns` an implementation detail of `TextFormatterImpl` by removing them from `TextLineBreak`. A new internal `WrappingTextLineBreak` class is introduced instead. The remaining runs are cleared once they're processed instead of staying in memory for the whole `TextLine` lifetime and their containing list can be reused.

## Perf changes

Benchmarks with changes:

|                Method |  Wrap |  Trim |     Before |      After | Change |
|---------------------- |------ |------ |-----------:|-----------:|-------:|
|       BuildTextLayout |  True | False | 1,414.2 us |   534.7 us |   -62% |
| BuildEmojisTextLayout |  True | False | 7,039.2 us | 1,545.2 us |   -78% |
|       BuildTextLayout |  True |  True | 1,552.2 us |   580.3 us |   -63% |
| BuildEmojisTextLayout |  True |  True | 7,318.1 us | 1,576.7 us |   -78% |