@Gillibald 2023-01-24
The next step will be to reuse the ShapedRuns via some external cache. So we will be able to keep shaped data during multiple measure / arrange calls if the text source doesn't change but the constraint.

Signature of FormatLine will look like so: 
`TextLine FormatLine(ITextSource textSource, int firstTextSourceIndex, double paragraphWidth, TextParagraphProperties paragraphProperties, TextLineBreak? previousLineBreak = null, ShapedRunCache? shapedRunCache)` 

Shaped runs can be added to that cache and could be reused.