@MrJul 2023-01-24
That would be pretty nice! (The `TextLayout` itself could keep a cache of its previous runs.)

It would definitely improve raw performance while re-layouting. In addition to that, `ShapedRun` and `GlyphRun` are pretty large objects and can contribute to significant garbage collections when discarded.